v0.0.1 (#1) (#2)
* criação do crud para sales com events

* changes:
1. code to publish events to console
2. new update endpoint

* aplica regras para permitir update

* 1. Adjusting contract objects
2. Adding log to mediatr pipeline

* code enhancements

* adding unit tests and adjusting mapping profile for saleitems

* converting some rules to specification

* Improvement in sales validation and logic
Adds specifications to check if a sale can be updated or deleted.
Enhances item validation in UpdateSaleCommandValidator to allow null lists, ensuring that if not null, they contain at least one item.
Includes new tests to validate business rules and the implemented changes.

* Adds asynchronous validations and tests for sales
Asynchronous validations were implemented in the Sale and SaleItem classes with specific validators. The BaseEntity class was changed to be abstract, allowing for custom validation logic. The ValidateAsync method was added to check business rules.
Validation tests were implemented for CreateSaleCommand, DeleteSaleCommand, GetSaleByIdQuery, GetSalesQuery, UpdateSaleCommand, SaleItemValidator, and SaleValidator. The test code for CreateSaleHandler was updated to reflect the new validations, with the removal of commented tests and the addition of new cases.
The validation structure was improved with more descriptive error messages and checks for specific conditions, such as the number of items and the presence of required fields.

* Refactoring of events and database configuration
• Improved error message readability in the CreateSaleHandler class.
• Added using directives for System.Text.Json and changed serialization in ConsolePublisher.
• Added reference to the Microsoft.EntityFrameworkCore.InMemory package in the project.
• Changed the database configuration to use an in-memory database in Program.cs.
• Removed unused using directives for code cleanup.

* InfrastructureModuleInitializer.cs now includes support for IMemoryCache and IConfiguration, registering the sales repository as either CachedSaleRepository or SaleRepository based on the EnableSaleCache configuration.
A new Features section was added to appsettings.json to configure sales cache, including cache duration.
The new file CachedSaleRepository.cs implements the cache logic, managing data storage and invalidation, improving the efficiency of listing operations.

* Correction DI

* Refactoring of sales classes and validations
Changed the property name from SaleDate to Date in several classes for standardization. Added validations in command classes and validators, including minimum date check and valid status verification. Changed the response structure from simple records to classes with initialization logic and required properties. Updated the mapping between entities and DTOs, including the change from SaleItemDto to SaleDto. Disabled sales cache in appsettings.json. Tests updated to reflect the new properties and validations.

* cleaning up and correcting language

* README.md

* adjusting api contract to adhere to specifications

* correcting swagger docs

* adjusting validators, validation pipeline and validationExceptionMiddleware to set default responses